+++ src/client/utils/test_utils/mod.rs
@@ -52,7 +52,10 @@ where
     for duration in &backoff {
         match f().await {
             Ok(val) => return Ok(val),
-            Err(_) => tokio::time::sleep(duration).await,
+            Err(_) => {
+                tokio::time::sleep(duration).await;
+                debug!("**** retrying *** ");
+            }
         }
     }
 
@@ -60,8 +63,8 @@ where
 }
 
 fn get_backoff_policy(retries: u8) -> Backoff {
-    let min = Duration::from_millis(100);
-    let max = Duration::from_secs(10);
+    let min = Duration::from_millis(500);
+    let max = Duration::from_secs(30);
     Backoff::new(retries as u32, min, max)
 }
 
@@ -109,7 +112,7 @@ macro_rules! retry_loop {
         loop {
             match $async_func.await {
                 Ok(val) => break val,
-                Err(_) => tokio::time::sleep(std::time::Duration::from_millis(200)).await,
+                Err(_) => tokio::time::sleep(std::time::Duration::from_secs(2)).await,
             }
         }
     };
@@ -122,7 +125,7 @@ macro_rules! retry_err_loop {
     ($async_func:expr) => {
         loop {
             match $async_func.await {
-                Ok(_) => tokio::time::sleep(std::time::Duration::from_millis(200)).await,
+                Ok(_) => tokio::time::sleep(std::time::Duration::from_secs(2)).await,
                 Err(err) => break err,
             }
         }
@@ -138,7 +141,7 @@ macro_rules! retry_loop_for_pattern {
             let result = $async_func.await;
             match &result {
                 $pattern $(if $cond)? => break result,
-                Ok(_) | Err(_) => tokio::time::sleep(std::time::Duration::from_millis(200)).await,
+                Ok(_) | Err(_) => tokio::time::sleep(std::time::Duration::from_secs(2)).await,
             }
         }
     };